Search a place and draw in a map
