+++ src/components/Map.jsx
@@ -7,6 +7,7 @@ class Map extends Component {
   }
 
   componentDidMount() {
+    console.log(this.props.destination);
     const map = new window.google.maps.Map(
       document.getElementById(this.props.id),
       this.props.options);
@@ -31,7 +32,7 @@ class Map extends Component {
   calculateAndDisplayRoute(directionsService, directionsRenderer) {
     directionsService.route({
       origin: {lat: this.props.position.lat, lng: this.props.position.lng},
-      destination: {lat: this.props.destination.lat, lng: this.props.destination.lng},
+      destination: {query: this.props.destination},
       travelMode: google.maps.TravelMode['DRIVING']
     }, function(response, status) {
       if (status == 'OK') {
