+++ src/index.js
@@ -8,10 +8,10 @@ import App from './routes/App';
 const initialState = {
     "user": {},
     "position": {},
-    "destination": {},
+    "destination": "",
     "map": {},
     "favorites": {},
-    "Record": {}
+    "Record": []
 };
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
